requirements.txt
yfinance>=0.2.36
ccxt>=4.2.14
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
transformers>=4.36.0
torch>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
